pyproject.toml
[project]
name = "daily-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-allauth>=65.13.1",
    "django-anymail[sendinblue]>=14.0",
    "google-generativeai>=0.8.6",
    "newsapi-python>=0.2.7",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
